mobile_app/Backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
pydantic[email]
bcrypt==4.0.1
passlib[bcrypt]==1.7.4
PyJWT
python-jose[cryptography]
python-dotenv
email-validator
psycopg2-binary
